json: treat integers as long long
The json parser stored integers in long int, preventing backtraces with
64-bit addresses to be correctly read on 32-bit machines.

Closes #78.

+++ lib/json.h
@@ -70,7 +70,7 @@ struct sr_json_value
    union
    {
       int boolean;
-      long integer;
+      long long integer;
       double dbl;
 
       struct
